fix(router): make addRoute split list of routes into multiple routeconfigs with single route each
