+++ src/util.js
@@ -42,5 +42,21 @@ export function _resolveUrl(fragment, baseUrl, hasPushState) {
   return _createRootedPath(fragment, baseUrl, hasPushState);
 }
 
+export function _ensureArrayWithSingleRoutePerConfig(config) {
+  let routeConfigs = [];
+
+  if (Array.isArray(config.route)) {
+    for (let i = 0, ii = config.route.length; i < ii; ++i) {
+      let current = Object.assign({}, config);
+      current.route = config.route[i];
+      routeConfigs.push(current);
+    }
+  } else {
+    routeConfigs.push(Object.assign({}, config));
+  }
+
+  return routeConfigs;
+}
+
 const isRootedPath = /^#?\//;
 const isAbsoluteUrl = /^([a-z][a-z0-9+\-.]*:)?\/\//i;
